+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Bugfix) Propagate Lifecycle Mount
 - (Feature) PVC Member Status info
 - (Feature) Respect ToBeCleanedServers in Agency
+- (Improvement) Unify K8S Error Handling
 
 ## [1.2.19](https://github.com/arangodb/kube-arangodb/tree/1.2.19) (2022-10-05)
 - (Bugfix) Prevent changes when UID is wrong